Refactor and remove unused components in the codebase
