--- src/app/components/govuk/govuk-checkboxes-new/govuk-checkboxes-new.component.html
@@ -1,19 +1,0 @@
-<div class="govuk-form-group">
-  <fieldset class="govuk-fieldset" [id]="fieldSetId" [attr.aria-describedby]="fieldSetId ? fieldSetId + 'Hint' : ''">
-    <legend
-      class="govuk-fieldset__legend {{ legendClasses }}"
-      [attr.describedby]="legendHint ? fieldSetId + 'Hint' : ''"
-    >
-      {{ legendText }}
-    </legend>
-    @if (legendHint) {
-      <div [id]="fieldSetId ? fieldSetId + 'Hint' : ''" class="govuk-hint">
-        {{ legendHint }}
-      </div>
-    }
-
-    <div class="govuk-checkboxes {{ checkboxClasses }}" data-module="govuk-checkboxes">
-      <ng-content></ng-content>
-    </div>
-  </fieldset>
-</div>
