+++ src/app/components/govuk/govuk-checkboxes/govuk-checkboxes.component.html
@@ -1,5 +1,5 @@
-<div class="govuk-form-group" [formGroup]="getGroup">
-  <fieldset class="govuk-fieldset" [id]="fieldSetId">
+<div class="govuk-form-group">
+  <fieldset class="govuk-fieldset" [id]="fieldSetId" [attr.aria-describedby]="fieldSetId ? fieldSetId + 'Hint' : ''">
     <legend
       class="govuk-fieldset__legend {{ legendClasses }}"
       [attr.describedby]="legendHint ? fieldSetId + 'Hint' : ''"
@@ -11,49 +11,9 @@
         {{ legendHint }}
       </div>
     }
+
     <div class="govuk-checkboxes {{ checkboxClasses }}" data-module="govuk-checkboxes">
-      @for (input of checkboxInputs; track input.inputId) {
-        @if (input.inputTextDivider) {
-          <div [id]="input.inputId ? input.inputId + 'CheckboxesDivider' : ''" class="govuk-checkboxes__divider">
-            {{ input.inputTextDivider }}
-          </div>
-        }
-
-        <div class="govuk-checkboxes__item">
-          <input
-            class="govuk-checkboxes__input {{ input.inputClasses }}"
-            [id]="input.inputId"
-            [name]="input.inputName"
-            type="checkbox"
-            [value]="input.inputValue"
-            formControlName="{{ input.inputId }}"
-            (change)="handleToggleConditional(input.inputId)"
-          />
-          <label class="govuk-label govuk-checkboxes__label" [for]="input.inputId"> {{ input.inputLabel }} </label>
-          @if (input.inputHint) {
-            <div [id]="input.inputId ? input.inputId + 'ItemHint' : ''" class="govuk-hint govuk-checkboxes__hint">
-              {{ input.inputHint }}
-            </div>
-          }
-        </div>
-
-        @if (input?.conditional && toggleConditional[input.inputId]) {
-          <div class="govuk-checkboxes__conditional" [id]="input.inputId + 'Conditional'">
-            <div class="govuk-form-group">
-              <label class="govuk-label" [for]="input?.conditional?.inputId"
-                >{{ input?.conditional?.inputLabel }}
-              </label>
-              <input
-                class="govuk-input  {{ input?.conditional?.inputClasses }}"
-                [id]="input?.conditional?.inputId"
-                [name]="input?.conditional?.inputName"
-                type="text"
-                formControlName="{{ input?.conditional?.inputId }}"
-              />
-            </div>
-          </div>
-        }
-      }
+      <ng-content></ng-content>
     </div>
   </fieldset>
 </div>
